Split legacy MeshInterface test monolith
